requirements.txt
Django>=4.0.3
djangorestframework>=3.13.1
django-filter>=21.1
django-otp>=1.1.3
Markdown>=3.3.6
qrcode>=7.3.1
requests>=2.27.1
google-cloud-compute>=1.0.0
google-cloud-secret-manager>=2.8.0
azure-identity>=1.7.1
azure-mgmt-compute>=25.0.0
boto3>=1.21.3